+++ src/main/java/com/extendedae_plus/mixin/ae2/accessor/AEBaseScreenAccessor.java
@@ -1,11 +1,10 @@
-package com.extendedae_plus.mixin.accessor;
-
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.gen.Accessor;
+package com.extendedae_plus.mixin.ae2.accessor;
 
 import appeng.client.gui.AEBaseScreen;
 import appeng.client.gui.style.ScreenStyle;
 import appeng.menu.AEBaseMenu;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
 
 @Mixin(value = AEBaseScreen.class, remap = false)
 public interface AEBaseScreenAccessor<T extends AEBaseMenu> {
