+++ src/main/java/com/extendedae_plus/mixin/ae2/accessor/MEStorageScreenAccessor.java
@@ -1,11 +1,10 @@
-package com.extendedae_plus.mixin.accessor;
-
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.gen.Accessor;
-import org.spongepowered.asm.mixin.gen.Invoker;
+package com.extendedae_plus.mixin.ae2.accessor;
 
 import appeng.client.gui.me.common.MEStorageScreen;
 import appeng.client.gui.widgets.AETextField;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+import org.spongepowered.asm.mixin.gen.Invoker;
 
 @Mixin(value = MEStorageScreen.class, remap = false)
 public interface MEStorageScreenAccessor {
