+++ src/main/java/com/extendedae_plus/mixin/ae2/accessor/PatternProviderLogicAccessor.java
@@ -1,10 +1,9 @@
-package com.extendedae_plus.mixin.accessor;
-
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.gen.Accessor;
+package com.extendedae_plus.mixin.ae2.accessor;
 
 import appeng.helpers.patternprovider.PatternProviderLogic;
 import appeng.helpers.patternprovider.PatternProviderLogicHost;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
 
 @Mixin(PatternProviderLogic.class)
 public interface PatternProviderLogicAccessor {
