+++ src/main/java/com/extendedae_plus/mixin/ae2/accessor/PatternProviderLogicPatternInputsAccessor.java
@@ -1,12 +1,11 @@
-package com.extendedae_plus.mixin.accessor;
+package com.extendedae_plus.mixin.ae2.accessor;
 
+import appeng.api.stacks.AEKey;
+import appeng.helpers.patternprovider.PatternProviderLogic;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
 
 import java.util.Set;
-
-import appeng.api.stacks.AEKey;
-import appeng.helpers.patternprovider.PatternProviderLogic;
 
 @Mixin(PatternProviderLogic.class)
 public interface PatternProviderLogicPatternInputsAccessor {
