+++ src/main/java/com/extendedae_plus/mixin/ae2WTlib/ContainerUWirelessExPatternTerminalMixin.java
@@ -1,9 +1,8 @@
-package com.extendedae_plus.mixin;
+package com.extendedae_plus.mixin.ae2WTlib;
 
-import appeng.api.util.IConfigurableObject;
 import com.extendedae_plus.util.ExtendedAEPatternUploadUtil;
-import com.glodblock.github.extendedae.container.ContainerWirelessExPAT;
-import com.glodblock.github.extendedae.common.me.itemhost.HostWirelessExPAT;
+import com.glodblock.github.extendedae.xmod.wt.ContainerUWirelessExPAT;
+import com.glodblock.github.extendedae.xmod.wt.HostUWirelessExPAT;
 import com.glodblock.github.glodium.network.packet.sync.IActionHolder;
 import com.glodblock.github.glodium.network.packet.sync.Paras;
 import net.minecraft.server.level.ServerPlayer;
@@ -20,11 +19,11 @@ import java.util.Map;
 import java.util.function.Consumer;
 
 /**
- * 为无线样板访问终端容器注册通用动作（CGenericPacket 分发）
+ * 为通用无线样板访问终端（AE2WTlib 集成）容器注册通用动作（CGenericPacket 分发）
  */
 @Pseudo
-@Mixin(ContainerWirelessExPAT.class)
-public abstract class ContainerWirelessExPatternTerminalMixin implements IActionHolder {
+@Mixin(ContainerUWirelessExPAT.class)
+public abstract class ContainerUWirelessExPatternTerminalMixin implements IActionHolder {
 
     @Unique
     private final Map<String, Consumer<Paras>> actions = createHolder();
@@ -32,9 +31,9 @@ public abstract class ContainerWirelessExPatternTerminalMixin implements IAction
     @Unique
     private Player epp$player;
 
-    // 明确目标构造签名：<init>(int, Inventory, HostWirelessExPAT)
-    @Inject(method = "<init>(ILnet/minecraft/world/entity/player/Inventory;Lcom/glodblock/github/extendedae/common/me/itemhost/HostWirelessExPAT;)V", at = @At("TAIL"), require = 0)
-    private void init(int id, net.minecraft.world.entity.player.Inventory playerInventory, HostWirelessExPAT host, CallbackInfo ci) {
+    // 明确目标构造签名：<init>(int, Inventory, HostUWirelessExPAT)
+    @Inject(method = "<init>(ILnet/minecraft/world/entity/player/Inventory;Lcom/glodblock/github/extendedae/xmod/wt/HostUWirelessExPAT;)V", at = @At("TAIL"), remap = false)
+    private void init(int id, net.minecraft.world.entity.player.Inventory playerInventory, HostUWirelessExPAT host, CallbackInfo ci) {
         this.epp$player = playerInventory.player;
         // 注册上传动作：参数顺序必须与客户端 CGenericPacket 保持一致
         this.actions.put("upload", p -> {
@@ -49,6 +48,7 @@ public abstract class ContainerWirelessExPatternTerminalMixin implements IAction
             }
         });
     }
+
 
     @NotNull
     @Override
