+++ src/main/java/com/extendedae_plus/mixin/extendedae/ContainerExPatternTerminalMixin.java
@@ -1,16 +1,15 @@
-package com.extendedae_plus.mixin;
+package com.extendedae_plus.mixin.extendedae;
 
 import appeng.api.util.IConfigurableObject;
+import appeng.menu.guisync.GuiSync;
 import com.extendedae_plus.util.ExtendedAEPatternUploadUtil;
-import com.glodblock.github.extendedae.xmod.wt.ContainerUWirelessExPAT;
-import com.glodblock.github.extendedae.xmod.wt.HostUWirelessExPAT;
+import com.glodblock.github.extendedae.container.ContainerExPatternTerminal;
 import com.glodblock.github.glodium.network.packet.sync.IActionHolder;
 import com.glodblock.github.glodium.network.packet.sync.Paras;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.player.Player;
 import org.jetbrains.annotations.NotNull;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Pseudo;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -19,12 +18,27 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import java.util.Map;
 import java.util.function.Consumer;
 
-/**
- * 为通用无线样板访问终端（AE2WTlib 集成）容器注册通用动作（CGenericPacket 分发）
- */
-@Pseudo
-@Mixin(ContainerUWirelessExPAT.class)
-public abstract class ContainerUWirelessExPatternTerminalMixin implements IActionHolder {
+@Mixin(ContainerExPatternTerminal.class)
+public abstract class ContainerExPatternTerminalMixin implements IActionHolder {
+
+    @GuiSync(11452)
+    @Unique
+    public boolean hidePatternSlots = false;
+
+    @Unique
+    public boolean isHidePatternSlots() {
+        return this.hidePatternSlots;
+    }
+
+    @Unique
+    public void setHidePatternSlots(boolean hide) {
+        this.hidePatternSlots = hide;
+    }
+
+    @Unique
+    public void toggleHidePatternSlots() {
+        this.hidePatternSlots = !this.hidePatternSlots;
+    }
 
     @Unique
     private final Map<String, Consumer<Paras>> actions = createHolder();
@@ -32,9 +46,8 @@ public abstract class ContainerUWirelessExPatternTerminalMixin implements IActio
     @Unique
     private Player epp$player;
 
-    // 明确目标构造签名：<init>(int, Inventory, HostUWirelessExPAT)
-    @Inject(method = "<init>(ILnet/minecraft/world/entity/player/Inventory;Lcom/glodblock/github/extendedae/xmod/wt/HostUWirelessExPAT;)V", at = @At("TAIL"), remap = false)
-    private void init(int id, net.minecraft.world.entity.player.Inventory playerInventory, HostUWirelessExPAT host, CallbackInfo ci) {
+    @Inject(method = "<init>", at = @At("TAIL"))
+    private void init(int id, net.minecraft.world.entity.player.Inventory playerInventory, IConfigurableObject host, CallbackInfo ci) {
         this.epp$player = playerInventory.player;
         // 注册上传动作：参数顺序必须与客户端 CGenericPacket 保持一致
         this.actions.put("upload", p -> {
@@ -49,7 +62,6 @@ public abstract class ContainerUWirelessExPatternTerminalMixin implements IActio
             }
         });
     }
-
 
     @NotNull
     @Override
