+++ src/main/java/com/extendedae_plus/mixin/extendedae/ContainerWirelessExPatternTerminalMixin.java
@@ -1,14 +1,15 @@
-package com.extendedae_plus.mixin;
+package com.extendedae_plus.mixin.extendedae;
 
-import appeng.menu.guisync.GuiSync;
 import com.extendedae_plus.util.ExtendedAEPatternUploadUtil;
-import appeng.api.util.IConfigurableObject;
-import com.glodblock.github.extendedae.container.ContainerExPatternTerminal;
+import com.glodblock.github.extendedae.common.me.itemhost.HostWirelessExPAT;
+import com.glodblock.github.extendedae.container.ContainerWirelessExPAT;
 import com.glodblock.github.glodium.network.packet.sync.IActionHolder;
 import com.glodblock.github.glodium.network.packet.sync.Paras;
 import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.entity.player.Player;
 import org.jetbrains.annotations.NotNull;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Pseudo;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -16,29 +17,13 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 import java.util.Map;
 import java.util.function.Consumer;
-import net.minecraft.world.entity.player.Player;
 
-@Mixin(ContainerExPatternTerminal.class)
-public abstract class ContainerExPatternTerminalMixin implements IActionHolder {
-
-    @GuiSync(11452)
-    @Unique
-    public boolean hidePatternSlots = false;
-
-    @Unique
-    public boolean isHidePatternSlots() {
-        return this.hidePatternSlots;
-    }
-
-    @Unique
-    public void setHidePatternSlots(boolean hide) {
-        this.hidePatternSlots = hide;
-    }
-
-    @Unique
-    public void toggleHidePatternSlots() {
-        this.hidePatternSlots = !this.hidePatternSlots;
-    }
+/**
+ * 为无线样板访问终端容器注册通用动作（CGenericPacket 分发）
+ */
+@Pseudo
+@Mixin(ContainerWirelessExPAT.class)
+public abstract class ContainerWirelessExPatternTerminalMixin implements IActionHolder {
 
     @Unique
     private final Map<String, Consumer<Paras>> actions = createHolder();
@@ -46,8 +31,9 @@ public abstract class ContainerExPatternTerminalMixin implements IActionHolder {
     @Unique
     private Player epp$player;
 
-    @Inject(method = "<init>", at = @At("TAIL"))
-    private void init(int id, net.minecraft.world.entity.player.Inventory playerInventory, IConfigurableObject host, CallbackInfo ci) {
+    // 明确目标构造签名：<init>(int, Inventory, HostWirelessExPAT)
+    @Inject(method = "<init>(ILnet/minecraft/world/entity/player/Inventory;Lcom/glodblock/github/extendedae/common/me/itemhost/HostWirelessExPAT;)V", at = @At("TAIL"), require = 0)
+    private void init(int id, net.minecraft.world.entity.player.Inventory playerInventory, HostWirelessExPAT host, CallbackInfo ci) {
         this.epp$player = playerInventory.player;
         // 注册上传动作：参数顺序必须与客户端 CGenericPacket 保持一致
         this.actions.put("upload", p -> {
