+++ src/main/resources/extendedae_plus.mixins.json
@@ -4,36 +4,39 @@
   "compatibilityLevel": "JAVA_17",
   "minVersion": "0.8",
   "client": [
-    "GuiExPatternProviderMixin",
-    "SlotGridLayoutMixin",
-    "GuiExPatternTerminalMixin",
-    "HighlightButtonMixin",
     "PickFromWirelessMixin",
-    "PatternEncodingTermScreenMixin",
-    "ae2.PatternProviderScreenMixin",
-    "jei.EncodePatternTransferHandlerMixin",
-    "ae2.QuartzCuttingKnifeItemMixin",
-    "accessor.AEBaseScreenAccessor",
     "accessor.AbstractContainerScreenAccessor",
     "accessor.ScreenAccessor",
-    "accessor.MEStorageScreenAccessor"
+    "accessor.ScreenInvoker",
+    "ae2.PatternEncodingTermScreenMixin",
+    "ae2.PatternProviderScreenMixin",
+    "ae2.QuartzCuttingKnifeItemMixin",
+    "ae2.SlotGridLayoutMixin",
+    "ae2.accessor.AEBaseScreenAccessor",
+    "ae2.accessor.AEBaseScreenInvoker",
+    "ae2.accessor.MEStorageScreenAccessor",
+    "extendedae.GuiExPatternProviderMixin",
+    "extendedae.GuiExPatternTerminalMixin",
+    "extendedae.HighlightButtonMixin",
+    "jei.EncodePatternTransferHandlerMixin"
   ],
   "mixins": [
-    "ContainerExPatternProviderMixin",
-    "ContainerExPatternTerminalMixin",
-    "ContainerWirelessExPatternTerminalMixin",
-    "ContainerUWirelessExPatternTerminalMixin",
-    "TileExPatternProviderMixin",
-    "PartExPatternProviderMixin",
-    "PatternEncodingTermMenuMixin",
-    "ContainerPatternEncodingTermMenuMixin",
-    "MEStorageMenuMixin",
-    "accessor.MEStorageMenuAccessor",
-    "accessor.PatternEncodingTermMenuAccessor",
+    "ae2.ContainerPatternEncodingTermMenuMixin",
+    "ae2.MEStorageMenuMixin",
+    "ae2.PatternEncodingTermMenuMixin",
     "ae2.PatternProviderLogicAdvancedMixin",
     "ae2.PatternProviderMenuAdvancedMixin",
-    "accessor.PatternProviderMenuAdvancedAccessor",
-    "accessor.PatternProviderLogicAccessor"
+    "ae2.accessor.MEStorageMenuAccessor",
+    "ae2.accessor.PatternEncodingTermMenuAccessor",
+    "ae2.accessor.PatternProviderLogicAccessor",
+    "ae2.accessor.PatternProviderLogicPatternInputsAccessor",
+    "ae2.accessor.PatternProviderMenuAdvancedAccessor",
+    "ae2WTlib.ContainerUWirelessExPatternTerminalMixin",
+    "extendedae.ContainerExPatternProviderMixin",
+    "extendedae.ContainerExPatternTerminalMixin",
+    "extendedae.ContainerWirelessExPatternTerminalMixin",
+    "extendedae.PartExPatternProviderMixin",
+    "extendedae.TileExPatternProviderMixin"
   ],
   "injectors": {
     "defaultRequire": 1
